+++ example/lib/state_restoration.dart
@@ -35,20 +35,20 @@ class _AppState extends State<App> with RestorationMixin {
     routes: [
       GoRoute(
         path: '/',
-        builder: (context, state) => MaterialPage<Page1Page>(
+        builder: (context, state) => MaterialPage<void>(
           key: state.pageKey,
           child: const Page1Page(),
         ),
       ),
       GoRoute(
         path: '/page2',
-        builder: (context, state) => MaterialPage<Page2Page>(
+        builder: (context, state) => MaterialPage<void>(
           key: state.pageKey,
           child: const Page2Page(),
         ),
       ),
     ],
-    error: (context, state) => MaterialPage<ErrorPage>(
+    error: (context, state) => MaterialPage<void>(
       key: state.pageKey,
       child: ErrorPage(state.error),
     ),