+++ example/lib/sub_routes.dart
@@ -21,7 +21,7 @@ class App extends StatelessWidget {
     routes: [
       GoRoute(
         path: '/',
-        builder: (context, state) => MaterialPage<HomePage>(
+        builder: (context, state) => MaterialPage<void>(
           key: state.pageKey,
           child: HomePage(families: Families.data),
         ),
@@ -31,7 +31,7 @@ class App extends StatelessWidget {
             builder: (context, state) {
               final family = Families.family(state.params['fid']!);
 
-              return MaterialPage<FamilyPage>(
+              return MaterialPage<void>(
                 key: state.pageKey,
                 child: FamilyPage(family: family),
               );
@@ -43,7 +43,7 @@ class App extends StatelessWidget {
                   final family = Families.family(state.params['fid']!);
                   final person = family.person(state.params['pid']!);
 
-                  return MaterialPage<PersonPage>(
+                  return MaterialPage<void>(
                     key: state.pageKey,
                     child: PersonPage(family: family, person: person),
                   );
@@ -54,7 +54,7 @@ class App extends StatelessWidget {
         ],
       ),
     ],
-    error: (context, state) => MaterialPage<ErrorPage>(
+    error: (context, state) => MaterialPage<void>(
       key: state.pageKey,
       child: ErrorPage(state.error),
     ),